pyproject.toml
[project]
name = "shrink-borders"
version = "0.1.0"
description = "Remove uniform color borders from images"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.0.0",
]

[project.scripts]
shrink-borders = "shrink_borders:main"
